HTML source asthetics
It is worth noting that I would likely be fired if this was for a
production environment given that my console turns almost completely
white as sqlalchemy's logging breezes by with requests to the database.
There is probably a way to opportunistically load the smaller support
tables (special, cat and sub cat).  The whole thing could be cached and
user specific data fed from a JS?func handler that is JUST their data.